+++ packages/api-lib/libs/es.js
@@ -80,6 +80,8 @@ async function prepare(index) {
     'type': 'object',
     properties: {
       'datetime': { type: 'date' },
+      'created': { type: 'date' },
+      'updated': { type: 'date' },
       'eo:cloud_cover': { type: 'float' },
       'eo:gsd': { type: 'float' },
       'eo:constellation': { type: 'keyword' },
@@ -176,6 +178,8 @@ async function _stream() {
         if (itemCollection) {
           const flatProperties =
             Object.assign({}, itemCollection.properties, data.properties)
+          flatProperties.created = new Date().toISOString()
+          flatProperties.updated = new Date().toISOString()
           esDataObject = Object.assign({}, esDataObject, { properties: flatProperties })
         } else {
           logger.error(`${data.id} has no collection`)
@@ -306,10 +310,9 @@ function buildQuery(parameters) {
 
 
   if (intersects) {
-    const { geometry } = intersects
     must.push({
       geo_shape: {
-        geometry: { shape: geometry }
+        geometry: { shape: intersects }
       }
     })
   }
@@ -373,28 +376,34 @@ function buildSort(parameters) {
 }
 
 function buildFieldsFilter(parameters) {
-  const id = 'id'
   const { fields } = parameters
-  const _sourceInclude = []
-  const _sourceExclude = []
+  let _sourceInclude = [
+    'id',
+    'type',
+    'geometry',
+    'bbox',
+    'links',
+    'assets',
+    'properties.datetime'
+  ]
+  let _sourceExclude = []
   if (fields) {
     const { include, exclude } = fields
-    if (include && include.length > 0) {
-      const propertiesIncludes = include.map(
-        (field) => (`${field}`)
-      ).concat(
-        [id]
-      )
-      _sourceInclude.push(...propertiesIncludes)
-    }
+    // Remove exclude fields from the default include list and add them to the source exclude list
     if (exclude && exclude.length > 0) {
-      const filteredExcludes = exclude.filter((field) =>
-        (![id].includes(field)))
-      const propertiesExclude = filteredExcludes.map((field) => (`${field}`))
-      _sourceExclude.push(...propertiesExclude)
+      _sourceInclude = _sourceInclude.filter((field) => !exclude.includes(field))
+      _sourceExclude = exclude
+    }
+    // Add include fields to the source include list if they're not already in it
+    if (include && include.length > 0) {
+      include.forEach((field) => {
+        if (_sourceInclude.indexOf(field) < 0) {
+          _sourceInclude.push(field)
+        }
+      })
     }
   }
-  return { _sourceExclude, _sourceInclude }
+  return { _sourceInclude, _sourceExclude }
 }
 
 async function search(parameters, index = '*', page = 1, limit = 10) {
@@ -419,7 +428,7 @@ async function search(parameters, index = '*', page = 1, limit = 10) {
     from: (page - 1) * limit
   }
 
-  const { _sourceExclude, _sourceInclude } = buildFieldsFilter(parameters)
+  const { _sourceInclude, _sourceExclude } = buildFieldsFilter(parameters)
   if (_sourceExclude.length > 0) {
     searchParams._sourceExclude = _sourceExclude
   }
@@ -431,10 +440,10 @@ async function search(parameters, index = '*', page = 1, limit = 10) {
   const results = resultBody.hits.hits.map((r) => (r._source))
   const response = {
     results,
-    meta: {
-      page,
+    'search:metadata': {
+      next: (((page * limit) < resultBody.hits.total) ? page + 1 : null),
       limit,
-      found: resultBody.hits.total,
+      matched: resultBody.hits.total,
       returned: results.length
     }
   }